# api-spec/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Optimus AI Platform API",
    "description": "Full REST API specification for the Optimus AI platform covering authentication, users, chats, models, files, knowledge, prompts, tools, groups, functions, skills, memories, notes, channels, automations, calendar, folders, configs, and third-party integrations.",
    "version": "1.0.0",
    "contact": {
      "name": "Optimus AI Support",
      "url": "https://github.com/optimus-ai"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development server"
    }
  ],
  "security": [
    { "BearerAuth": [] },
    { "ApiKeyAuth": [] }
  ],
...
  "paths": {
...
    "/api/v1/files/": {
      "get": {
        "tags": ["Files"],
        "summary": "List files",
        "operationId": "listFiles",
        "responses": {
          "200": { "description": "Files", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/FileObject" } } } } }
        }
      },
      "post": {
        "tags": ["Files"],
        "summary": "Upload file",
        "operationId": "uploadFile",
        "requestBody": {
          "required": true,
          "content": { "multipart/form-data": { "schema": { "type": "object", "properties": { "file": { "type": "string", "format": "binary" } }, "required": ["file"] } } }
        },
        "responses": {
          "200": { "description": "File uploaded", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/FileObject" } } } }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token. Obtained via /api/v1/auths/signin or /api/v1/auths/signup."
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "API key with 'sk-' prefix. Passed as 'Bearer sk-...'"
      }
    },
...
    "schemas": {
...
      "FileObject": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "path": { "type": "string" },
          "size": { "type": "integer" },
          "mimeType": { "type": "string" },
          "hash": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
    }
  }
}